refactor(audio): fold WebAudioDiagnostics + AudioPlayerDiagnostics into AudioDiagnosticsUI.cs
Both records live next to AudioDiagnosticsState now; their standalone files are
removed. AudioContextDiagnostics stays in its own file.
