--- src/dotnet/UI.Blazor.App/Services/AudioDiagnostics/AudioPlayerDiagnostics.cs
@@ -1,11 +1,0 @@
-namespace ActualChat.UI.Blazor.App.Services;
-
-public sealed record AudioPlayerDiagnostics(
-    string InternalId,
-    string? AuthorId,
-    string PlaybackState,
-    string BufferState,
-    double? PresentationLagMs,
-    double TargetBufferSizeMs,
-    double PlayingAt,
-    double BufferedDuration);
